+++ standingssync/utils.py
@@ -1,21 +1,24 @@
 import socket
-from datetime import timedelta
+from datetime import datetime, timedelta
+import json
 import logging
 import os
 import re
+from typing import Any
+from urllib.parse import urljoin
+
+from pytz import timezone
 
 from django.apps import apps
 from django.conf import settings
-from django.contrib.auth.models import User, Permission
 from django.contrib.messages.constants import DEBUG, ERROR, SUCCESS, WARNING, INFO
 from django.contrib import messages
-from django.db.models import Q
+from django.db import models
 from django.test import TestCase
+from django.urls import reverse
 from django.utils.functional import lazy
 from django.utils.html import format_html, mark_safe
 from django.utils.translation import gettext_lazy as _
-
-from allianceauth.notifications import notify
 
 
 # Format for output of datetime for this app
@@ -27,8 +30,8 @@ format_html_lazy = lazy(format_html, str)
 class LoggerAddTag(logging.LoggerAdapter):
     """add custom tag to a logger"""
 
-    def __init__(self, logger, prefix):
-        super(LoggerAddTag, self).__init__(logger, {})
+    def __init__(self, my_logger, prefix):
+        super(LoggerAddTag, self).__init__(my_logger, {})
         self.prefix = prefix
 
     def process(self, msg, kwargs):
@@ -138,22 +141,6 @@ class messages_plus:
         )
 
 
-def notify_admins(message: str, title: str, level="info") -> None:
-    """send notification to all admins"""
-    try:
-        perm = Permission.objects.get(codename="logging_notifications")
-        users = User.objects.filter(
-            Q(groups__permissions=perm)
-            | Q(user_permissions=perm)
-            | Q(is_superuser=True)
-        ).distinct()
-
-        for user in users:
-            notify(user, title=title, message=message, level=level)
-    except Permission.DoesNotExist:
-        pass
-
-
 def chunks(lst, size):
     """Yield successive sized chunks from lst."""
     for i in range(0, len(lst), size):
@@ -166,7 +153,8 @@ def clean_setting(
     min_value: int = None,
     max_value: int = None,
     required_type: type = None,
-):
+    choices: list = None,
+) -> Any:
     """cleans the input for a custom setting
 
     Will use `default_value` if settings does not exit or has the wrong type
@@ -175,6 +163,8 @@ def clean_setting(
     Need to define `required_type` if `default_value` is `None`
 
     Will assume `min_value` of 0 for int (can be overriden)
+
+    `None` allowed as value
 
     Returns cleaned value for setting
     """
@@ -190,12 +180,14 @@ def clean_setting(
     if not hasattr(settings, name):
         cleaned_value = default_value
     else:
-        if (
-            isinstance(getattr(settings, name), required_type)
-            and (min_value is None or getattr(settings, name) >= min_value)
-            and (max_value is None or getattr(settings, name) <= max_value)
+        dirty_value = getattr(settings, name)
+        if dirty_value is None or (
+            isinstance(dirty_value, required_type)
+            and (min_value is None or dirty_value >= min_value)
+            and (max_value is None or dirty_value <= max_value)
+            and (choices is None or dirty_value in choices)
         ):
-            cleaned_value = getattr(settings, name)
+            cleaned_value = dirty_value
         else:
             logger.warn(
                 "You setting for {} it not valid. Please correct it. "
@@ -222,11 +214,11 @@ def set_test_logger(logger_name: str, name: str) -> object:
     )
     f_handler = logging.FileHandler("{}.log".format(os.path.splitext(name)[0]), "w+")
     f_handler.setFormatter(f_format)
-    logger = logging.getLogger(logger_name)
-    logger.level = logging.DEBUG
-    logger.addHandler(f_handler)
-    logger.propagate = False
-    return logger
+    my_logger = logging.getLogger(logger_name)
+    my_logger.level = logging.DEBUG
+    my_logger.addHandler(f_handler)
+    my_logger.propagate = False
+    return my_logger
 
 
 def timeuntil_str(duration: timedelta) -> str:
@@ -304,22 +296,25 @@ def yesno_str(value: bool) -> str:
 
 def get_site_base_url() -> str:
     """return base URL for this site"""
-
-    base_url = "http://www.example.com"
-    if hasattr(settings, "ESI_SSO_CALLBACK_URL"):
+    try:
         match = re.match(r"(.+)\/sso\/callback", settings.ESI_SSO_CALLBACK_URL)
         if match:
-            base_url = match.group(1)
+            return match.group(1)
+    except AttributeError:
+        pass
 
-    return base_url
+    return ""
+
+
+def get_absolute_url(url_name: str) -> str:
+    """Returns absolute URL for the given URL name."""
+    return urljoin(get_site_base_url(), reverse(url_name))
 
 
 def dt_eveformat(dt: object) -> str:
     """converts a datetime to a string in eve format
     e.g. '2019-06-25T19:04:44'
     """
-    from datetime import datetime
-
     dt2 = datetime(dt.year, dt.month, dt.day, dt.hour, dt.minute, dt.second)
     return dt2.isoformat()
 
@@ -370,3 +365,70 @@ def create_bs_button_html(
         mark_safe(' disabled="disabled"') if disabled else "",
         create_bs_glyph_html(glyph_name),
     )
+
+
+def create_fa_button_html(
+    url: str,
+    fa_code: str,
+    button_type: str,
+    tooltip: str = None,
+    disabled: bool = False,
+) -> str:
+    """create BS botton and return HTML"""
+    return format_html(
+        '<a href="{}" class="btn btn-{}"{}>{}{}</a>',
+        url,
+        button_type,
+        mark_safe(f' title="{tooltip}"') if tooltip else "",
+        mark_safe(' disabled="disabled"') if disabled else "",
+        mark_safe(f'<i class="{fa_code}"></i>'),
+    )
+
+
+class JSONDateTimeDecoder(json.JSONDecoder):
+    def __init__(self, *args, **kwargs) -> None:
+        json.JSONDecoder.__init__(
+            self, object_hook=self.dict_to_object, *args, **kwargs
+        )
+
+    def dict_to_object(self, dct: dict) -> object:
+        if "__type__" not in dct:
+            return dct
+
+        type_str = dct.pop("__type__")
+        zone, _ = dct.pop("tz")
+        dct["tzinfo"] = timezone(zone)
+        try:
+            dateobj = datetime(**dct)
+            return dateobj
+        except (ValueError, TypeError):
+            dct["__type__"] = type_str
+            return dct
+
+
+class JSONDateTimeEncoder(json.JSONEncoder):
+    """Instead of letting the default encoder convert datetime to string,
+    convert datetime objects into a dict, which can be decoded by the
+    JSONDateTimeDecoder
+    """
+
+    def default(self, o: Any) -> Any:
+        if isinstance(o, datetime):
+            return {
+                "__type__": "datetime",
+                "year": o.year,
+                "month": o.month,
+                "day": o.day,
+                "hour": o.hour,
+                "minute": o.minute,
+                "second": o.second,
+                "microsecond": o.microsecond,
+                "tz": (o.tzinfo.tzname(o), o.utcoffset().total_seconds()),
+            }
+        else:
+            return json.JSONEncoder.default(self, o)
+
+
+def generate_invalid_pk(MyModel: models.Model) -> int:
+    """return an invalid PK for the given Django model"""
+    return MyModel.objects.aggregate(models.Max("pk"))["pk__max"] + 1